requirements.txt
# Core Framework
streamlit>=1.29.0

# TMDB API
requests>=2.31.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# NLP Libraries
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
textblob==0.17.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.3

# Machine Learning for Recommendations
scipy==1.11.4
surprise==0.1

# Utilities
python-dotenv==1.0.0
